+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 <!-- markdownlint-disable first-line-h1 -->
 ## v1.7.0
 
-*2026-06-10*
+*2026-07-07*
 
 **Changed**
 
 - [Color provider](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.18/specification/#textDocument_documentColor) is now supported by [culori](https://culorijs.org/) instead of unmaintained [color-rgba](https://www.npmjs.com/package/color-rgba)
+- Upgrade the [WikiLint](https://www.npmjs.com/package/wikilint) package to [v2.46.0](https://github.com/bhsd-harry/wikiparser-node/blob/main/CHANGELOG.md#v1460)
+- The document highlight provider now returns narrower ranges for extension tags, HTML tags and parameter keys
 
 ## v1.6.3